maybe this will please the windows gods

+++ packages/kit/src/utils/filesystem.js
@@ -13,6 +13,7 @@ export function mkdirp(dir) {
 
 /** @param {string} path */
 export function rimraf(path) {
+	if (!fs.existsSync(path)) return;
 	(fs.rmSync || fs.rmdirSync)(path, { recursive: true, force: true });
 }
